+++ example/lib/models/user.dart
@@ -3,13 +3,16 @@ import 'package:firestore_model/firestore_model.dart';
 class User extends FirestoreModel<User> {
   String? firstName;
   String? lastName;
+  List<String>? languages;
 
-  User({this.firstName, this.lastName});
+  User({this.firstName, this.lastName, this.languages});
 
   // use to read
   User.formMap(Map<String, dynamic> map) {
     this.firstName = map['first_name'];
     this.lastName = map['last_name'];
+    this.languages =
+        map['languages'] is List ? map['languages'].cast<String>() : [];
   }
 
   // use to write
@@ -17,6 +20,7 @@ class User extends FirestoreModel<User> {
   Map<String, dynamic> get toMap => {
         'first_name': this.firstName,
         'last_name': this.lastName,
+        'languages': this.languages
       };
 
   @override